// src/app/shared/core/cache/cache.enum.ts
export enum CommonCacheEnum {
    // Common cache enum values
    AssetType = 1,
    Counterparty = 2,
    Country = 3,
    Currency = 4,
    ISIN = 5,
    Theme = 6,
    TradeName = 7,
    TradeAccount = 8,
    Holidays = 9,
    RVUser = 10,
    ArchivingGroup = 11,
    FxOptionType = 12,
    CuttOffTime = 13,
    TradarAccount = 14,
    Indices = 15,
    DayCount = 16,
    FixingPage = 17,
    AssetTradeType = 18,
    BusinessDaysShifter = 19,
    CurrentAllocation = 20,
    AllocationTemplateFields = 21,
    TradeType = 22,
    TradeCategory = 23,
    UserTraders = 24,
    RVUserRole = 25,
    RVRole = 26,
    AllocationTemplate = 27,
    TradarAccountMapping = 28,
    ActiveTradeName = 29
}
export enum CommonCacheDictionaryEnum   {
    AssetType = "AssetType",
    Counterparty = "Counterparty",
    Country = "Country",
    Currency = "Currency",
    ISIN = "ISIN",
    Theme = "Theme",
    TradeName = "TradeName",
    TradeAccount = "TradeAccount",
    Holidays = "Holidays",
    RVUser = "RVUser",
    ArchivingGroup = "ArchivingGroup",
    FxOptionType = "FxOptionType",
    CuttOffTime = "CuttOffTime",
    TradarAccount = "TradarAccount",
    Indices = "Indices",
    DayCount = "DayCount",
    FixingPage = "FixingPage",
    AssetTradeType = "AssetTradeType",
    BusinessDaysShifter = "BusinessDaysShifter",
    CurrentAllocation = "CurrentAllocation",
    AllocationTemplateFields = "AllocationTemplateFields",
    TradeType = "TradeType",
    TradeCategory = "TradeCategory",
    UserTraders = "UserTraders",
    RVUserRole = "RVUserRole",
    RVRole = "RVRole",
    AllocationTemplate = "AllocationTemplate",
    TradarAccountMapping = "TradarAccountMapping",
    ActiveTradeName = "ActiveTradeName"
} 